embedding-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bufbuild-protovalidate-protocolbuffers-pyi==31.0.0.1.20250423154025+ec0af4932e5e
bufbuild-protovalidate-protocolbuffers-python==31.0.0.1.20250423154025+ec0af4932e5e
cel-python==0.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
clip-anytorch==2.6.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.3.1
grpcio==1.71.0
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.31.2
idna==3.10
inspire_protobuf @ git+https://github.com/tech-inspire/api-contracts@11fa777355f0267caeb544a0bf55a53dfc6aedcc
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
lark==0.12.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
nats-py==2.10.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.0
protovalidate==0.8.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
types-protobuf==6.30.2.20250516
types-python-dateutil==2.9.0.20250516
types-PyYAML==6.0.12.20250516
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
yarl==1.20.0
